change assetId to a long and add most of the client data tables and controllers

+++ grails-app/domain/dashboard/fuelaccounts/Asset.groovy
@@ -3,7 +3,7 @@ package dashboard.fuelaccounts
 class Asset {
 
     static belongsTo = [account: Accounts]
-    int assetId
+    long assetId
     String barcode
     int tankNum
